+++ TShockAPI/Hooks/PlayerHooks.cs
@@ -143,7 +143,7 @@ namespace TShockAPI.Hooks
 	/// <summary>
 	/// EventArgs used for the <see cref="PlayerHooks.PlayerPermission"/> event.
 	/// </summary>
-	public class PlayerPermissionEventArgs : HandledEventArgs
+	public class PlayerPermissionEventArgs
 	{
 		/// <summary>
 		/// The player who fired the event.
@@ -156,6 +156,11 @@ namespace TShockAPI.Hooks
 		public string Permission { get; set; }
 
 		/// <summary>
+		/// <see cref="PermissionHookResult"/> of the hook.
+		/// </summary>
+		public PermissionHookResult Result { get; set; }
+
+		/// <summary>
 		/// Initializes a new instance of the PlayerPermissionEventArgs class.
 		/// </summary>
 		/// <param name="player">The player who fired the event.</param>
@@ -164,13 +169,14 @@ namespace TShockAPI.Hooks
 		{
 			Player = player;
 			Permission = permission;
+			Result = PermissionHookResult.Unhandled;
 		}
 	}
 
 	/// <summary>
 	/// EventArgs used for the <see cref="PlayerHooks.PlayerItembanPermission"/> event.
 	/// </summary>
-	public class PlayerItembanPermissionEventArgs : HandledEventArgs
+	public class PlayerItembanPermissionEventArgs
 	{
 		/// <summary>
 		/// The player who fired the event.
@@ -183,21 +189,27 @@ namespace TShockAPI.Hooks
 		public ItemBan BannedItem { get; set; }
 
 		/// <summary>
+		/// <see cref="PermissionHookResult"/> of the hook.
+		/// </summary>
+		public PermissionHookResult Result { get; set; }
+
+		/// <summary>
 		/// Initializes a new instance of the PlayerItembanPermissionEventArgs class.
 		/// </summary>
 		/// <param name="player">The player who fired the event.</param>
-		/// <param name="permission">The permission being checked.</param>
+		/// <param name="bannedItem">The banned item being checked.</param>
 		public PlayerItembanPermissionEventArgs(TSPlayer player, ItemBan bannedItem)
 		{
 			Player = player;
 			BannedItem = bannedItem;
+			Result = PermissionHookResult.Unhandled;
 		}
 	}
 
 	/// <summary>
 	/// EventArgs used for the <see cref="PlayerHooks.PlayerProjbanPermission"/> event.
 	/// </summary>
-	public class PlayerProjbanPermissionEventArgs : HandledEventArgs
+	public class PlayerProjbanPermissionEventArgs
 	{
 		/// <summary>
 		/// The player who fired the event.
@@ -210,6 +222,11 @@ namespace TShockAPI.Hooks
 		public ProjectileBan BannedProjectile { get; set; }
 
 		/// <summary>
+		/// <see cref="PermissionHookResult"/> of the hook.
+		/// </summary>
+		public PermissionHookResult Result { get; set; }
+
+		/// <summary>
 		/// Initializes a new instance of the PlayerProjbanPermissionEventArgs class.
 		/// </summary>
 		/// <param name="player">The player who fired the event.</param>
@@ -218,13 +235,14 @@ namespace TShockAPI.Hooks
 		{
 			Player = player;
 			BannedProjectile = checkedProjectile;
+			Result = PermissionHookResult.Unhandled;
 		}
 	}
 
 	/// <summary>
 	/// EventArgs used for the <see cref="PlayerHooks.PlayerTilebanPermission"/> event.
 	/// </summary>
-	public class PlayerTilebanPermissionEventArgs : HandledEventArgs
+	public class PlayerTilebanPermissionEventArgs
 	{
 		/// <summary>
 		/// The player who fired the event.
@@ -237,6 +255,11 @@ namespace TShockAPI.Hooks
 		public TileBan BannedTile { get; set; }
 
 		/// <summary>
+		/// <see cref="PermissionHookResult"/> of the hook.
+		/// </summary>
+		public PermissionHookResult Result { get; set; }
+
+		/// <summary>
 		/// Initializes a new instance of the PlayerTilebanPermissionEventArgs class.
 		/// </summary>
 		/// <param name="player">The player who fired the event.</param>
@@ -245,6 +268,7 @@ namespace TShockAPI.Hooks
 		{
 			Player = player;
 			BannedTile = checkedTile;
+			Result = PermissionHookResult.Unhandled;
 		}
 	}
 
@@ -439,60 +463,79 @@ namespace TShockAPI.Hooks
 		/// Fires the <see cref="PlayerPermission"/> event.
 		/// </summary>
 		/// <param name="player">The player firing the event.</param>
-		/// <returns>True if the event has been handled.</returns>
-		public static bool OnPlayerPermission(TSPlayer player, string permission)
+		/// <returns>Event result if the event has been handled, otherwise <see cref="PermissionHookResult.Unhandled"/>.</returns>
+		public static PermissionHookResult OnPlayerPermission(TSPlayer player, string permission)
 		{
 			if (PlayerPermission == null)
-				return false;
+				return PermissionHookResult.Unhandled;
 
 			var args = new PlayerPermissionEventArgs(player, permission);
 			PlayerPermission(args);
-			return args.Handled;
+
+			return args.Result;
 		}
 
 		/// <summary>
 		/// Fires the <see cref="PlayerItembanPermission"/> event.
 		/// </summary>
 		/// <param name="player">The player firing the event.</param>
-		/// <returns>True if the event has been handled.</returns>
-		public static bool OnPlayerItembanPermission(TSPlayer player, ItemBan bannedItem)
+		/// <returns>Event result if the event has been handled, otherwise <see cref="PermissionHookResult.Unhandled"/>.</returns>
+		public static PermissionHookResult OnPlayerItembanPermission(TSPlayer player, ItemBan bannedItem)
 		{
 			if (PlayerItembanPermission == null)
-				return false;
+				return PermissionHookResult.Unhandled;
 
 			var args = new PlayerItembanPermissionEventArgs(player, bannedItem);
 			PlayerItembanPermission(args);
-			return args.Handled;
+
+			return args.Result;
 		}
 
 		/// <summary>
 		/// Fires the <see cref="PlayerProjbanPermission"/> event.
 		/// </summary>
 		/// <param name="player">The player firing the event.</param>
-		/// <returns>True if the event has been handled.</returns>
-		public static bool OnPlayerProjbanPermission(TSPlayer player, ProjectileBan bannedProj)
+		/// <returns>Event result if the event has been handled, otherwise <see cref="PermissionHookResult.Unhandled"/>.</returns>
+		public static PermissionHookResult OnPlayerProjbanPermission(TSPlayer player, ProjectileBan bannedProj)
 		{
 			if (PlayerProjbanPermission == null)
-				return false;
+				return PermissionHookResult.Unhandled;
 
 			var args = new PlayerProjbanPermissionEventArgs(player, bannedProj);
 			PlayerProjbanPermission(args);
-			return args.Handled;
+
+			return args.Result;
 		}
 
 		/// <summary>
 		/// Fires the <see cref="PlayerTilebanPermission"/> event.
 		/// </summary>
 		/// <param name="player">The player firing the event.</param>
-		/// <returns>True if the event has been handled.</returns>
-		public static bool OnPlayerTilebanPermission(TSPlayer player, TileBan bannedTile)
+		/// <returns>Event result if the event has been handled, otherwise <see cref="PermissionHookResult.Unhandled"/>.</returns>
+		public static PermissionHookResult OnPlayerTilebanPermission(TSPlayer player, TileBan bannedTile)
 		{
 			if (PlayerTilebanPermission == null)
-				return false;
+				return PermissionHookResult.Unhandled;
 
 			var args = new PlayerTilebanPermissionEventArgs(player, bannedTile);
 			PlayerTilebanPermission(args);
-			return args.Handled;
+
+			return args.Result;
 		}
+
 	}
+
+	/// <summary>
+	/// Defines the possible outcomes of <see cref="PlayerHooks.PlayerPermission"/> handlers.
+	/// </summary>
+	public enum PermissionHookResult
+	{
+		/// <summary>Hook doesn't return a result on the permission check.</summary>
+		Unhandled,
+		/// <summary>Permission is explicitly denied by a hook.</summary>
+		Denied,
+		/// <summary>Permission is explicitly granted by a hook.</summary>
+		Granted
+	}
+
 }